refactor: remove unnecessary interfaces and merge folders
