+++ profanity-domain/src/main/java/app/application/event/FilterEventHandler.java
@@ -1,6 +1,5 @@
 package app.application.event;
 
-import app.application.tracking.DefaultTrackingRecorder;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.stereotype.Component;
@@ -13,16 +12,15 @@ import static org.springframework.transaction.annotation.Propagation.REQUIRES_NE
 public class FilterEventHandler {
 
     private static final Logger log = LogManager.getLogger(FilterEventHandler.class);
-    private final DefaultTrackingRecorder trackingRecorder;
+    private final TrackingRecorder trackingRecorder;
 
-    public FilterEventHandler(DefaultTrackingRecorder trackingRecorder) {
+    public FilterEventHandler(TrackingRecorder trackingRecorder) {
         this.trackingRecorder = trackingRecorder;
     }
 
     @Transactional(propagation = REQUIRES_NEW)
     @TransactionalEventListener
     public void handle(FilterEvent event) {
-
         trackingRecorder.recordTracking(event);
         log.info("[DOMAIN] 이벤트 발행 성공 : event={}", event);
     }
