--- profanity-domain/src/main/java/app/application/tracking/TrackingRecorder.java
@@ -1,9 +1,0 @@
-package app.application.tracking;
-
-import app.application.event.FilterEvent;
-
-public interface TrackingRecorder {
-
-    void recordTracking(FilterEvent event);
-
-}
